Layout logic now renders to fractional DP values rather than ints.

+++ rover/src/main/kotlin/io/rover/rover/ui/types/Layout.kt
@@ -5,5 +5,9 @@ package io.rover.rover.ui.types
  */
 data class Layout(
     val coordinatesAndViewModels: CoordinatesAndViewModels,
-    val height: Int
+
+    /**
+     * The total height (in dp) of the entire layout from start to finish.
+     */
+    val height: Float
 )
